Change ValidationForm to MainForm
